1月10日：测试app，修改一些bug、界面，增加一些需求
1月10日：测试app，修改一些bug、界面，增加一些需求

+++ NZNYiOS/我的/Controller/CYMineViewController.m
@@ -173,10 +173,10 @@
     // 假数据
     CYMineVCModel *mineVCModel = self.allInfoArr[0];
     
-    NSString *fID = [NSString stringWithFormat:@"%ld",mineVCModel.FId];
+    NSString *fID = [NSString stringWithFormat:@"%ld",(long)mineVCModel.FId];
     NSString *money = [NSString stringWithFormat:@"¥ %.2f",mineVCModel.Money];
     NSString *certificateLevel = [NSString stringWithFormat:@"%.1f",mineVCModel.CertificateLevel];
-    NSString *like = [NSString stringWithFormat:@"%ld",mineVCModel.LikeCount];
+    NSString *like = [NSString stringWithFormat:@"%ld",(long)mineVCModel.LikeCount];
     
     
     // headerInfoModel
